chore: refactor to it's own module
